# contracts/openapi/breakglass.yaml
openapi: 3.1.0
info:
  title: PbD-SPMIS Break-Glass Manager
  version: 0.2.1
  x-privacy-headers:
  - X-Purpose
  - X-Program
  - X-Correlation-ID
  - X-Channel
  - X-Device-Trust
  - X-Case-ID
  - X-Break-Glass-Grant
  x-response-headers:
  - X-Decision-ID
  - X-Policy-Version
  - X-Obligations
  x-source: generated by scripts/export_contracts.py; do not edit by hand
paths:
  /healthz:
    get:
      summary: Healthz
      operationId: healthz_healthz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Healthz Healthz Get
  /v1/grants:
    post:
      summary: Request Grant
      operationId: request_grant_v1_grants_post
      parameters:
      - name: X-Correlation-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Correlation-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Request Grant V1 Grants Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: List Grants
      operationId: list_grants_v1_grants_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: pending|active|expired|revoked|reviewed|overdue_review
          title: Status
        description: pending|active|expired|revoked|reviewed|overdue_review
      - name: X-Correlation-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Correlation-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Grants V1 Grants Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/grants/{grant_id}/approve:
    post:
      summary: Approve
      operationId: approve_v1_grants__grant_id__approve_post
      parameters:
      - name: grant_id
        in: path
        required: true
        schema:
          type: string
          title: Grant Id
      - name: X-Correlation-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Correlation-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Approve V1 Grants  Grant Id  Approve Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/grants/{grant_id}/revoke:
    post:
      summary: Revoke
      operationId: revoke_v1_grants__grant_id__revoke_post
      parameters:
      - name: grant_id
        in: path
        required: true
        schema:
          type: string
          title: Grant Id
      - name: X-Correlation-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Correlation-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Revoke V1 Grants  Grant Id  Revoke Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/grants/{grant_id}/review:
    post:
      summary: Review
      operationId: review_v1_grants__grant_id__review_post
      parameters:
      - name: grant_id
        in: path
        required: true
        schema:
          type: string
          title: Grant Id
      - name: X-Correlation-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Correlation-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Review V1 Grants  Grant Id  Review Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/grants/{grant_id}:
    get:
      summary: Get Grant
      operationId: get_grant_v1_grants__grant_id__get
      parameters:
      - name: grant_id
        in: path
        required: true
        schema:
          type: string
          title: Grant Id
      - name: X-Correlation-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Correlation-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Grant V1 Grants  Grant Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    GrantIn:
      properties:
        subject_token:
          type: string
          title: Subject Token
        program:
          type: string
          title: Program
        attributes:
          items:
            type: string
          type: array
          minItems: 1
          title: Attributes
        reason_code:
          type: string
          pattern: ^[a-z_]+$
          title: Reason Code
          description: e.g. immediate_risk, court_order
        reason:
          type: string
          maxLength: 500
          minLength: 10
          title: Reason
        case_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Case Ref
        duration_minutes:
          type: integer
          maximum: 240.0
          minimum: 5.0
          title: Duration Minutes
          default: 60
      type: object
      required:
      - subject_token
      - program
      - attributes
      - reason_code
      - reason
      title: GrantIn
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ReviewIn:
      properties:
        outcome:
          type: string
          pattern: ^(justified|unjustified|needs_follow_up)$
          title: Outcome
        notes:
          type: string
          maxLength: 500
          title: Notes
          default: ''
      type: object
      required:
      - outcome
      title: ReviewIn
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Service or user token. Development tokens are HS256; production
        uses OIDC.
  parameters:
    XPurpose:
      name: X-Purpose
      in: header
      required: false
      schema:
        type: string
      description: Registered purpose (catalog/purposes.yaml). Required on sensitive
        endpoints.
    XProgram:
      name: X-Program
      in: header
      required: false
      schema:
        type: string
      description: Program the request is made for (see catalog/programs.yaml).
    XCorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
      description: Correlation id propagated to downstream calls and audit events.
    XChannel:
      name: X-Channel
      in: header
      required: false
      schema:
        type: string
      description: Access channel (api, portal, assisted, mobile).
    XDeviceTrust:
      name: X-Device-Trust
      in: header
      required: false
      schema:
        type: string
      description: 'Device/session trust asserted by the access layer: low, managed
        or high.'
    XCaseID:
      name: X-Case-ID
      in: header
      required: false
      schema:
        type: string
      description: Case the actor is working on; required by case-scoped purposes.
    XBreakGlassGrant:
      name: X-Break-Glass-Grant
      in: header
      required: false
      schema:
        type: string
      description: Active break-glass grant id for the emergency_protection purpose.
  headers:
    X-Decision-ID:
      schema:
        type: string
      description: Policy decision id recorded in the audit store.
    X-Policy-Version:
      schema:
        type: string
      description: Catalogue/policy version used for the decision.
    X-Obligations:
      schema:
        type: string
      description: Comma-separated obligations the caller must honour, e.g. no_export.
security:
- bearerAuth: []